added bable to test in package.json and separated server logic and app startup logic
